refactor(core): extract run planning into execute_plan.ts
Move the pre-execution planning decisions out of execute(): parameter
resolution with its secret registration, the interactive prompt fallback, the
condition-driven skip pruning, the dangling ordering-edge report, and the
--affected narrowing now live in a sibling module. Every comment moves with
the code it documents. No behaviour and no public API change.
