+++ packages/core/src/executor.ts
@@ -22,6 +22,13 @@ import {
   emitActionsMasks,
   writeJobSummary,
 } from "./execute_output.ts";
+import {
+  applyAffectedSkips,
+  conditionSkips,
+  defaultPrompt,
+  reportDanglingEdges,
+  resolveRunParameters,
+} from "./execute_plan.ts";
 import { makeLifecycle } from "./lifecycle.ts";
 import type { RunEnv, RunOutcome } from "./run_support.ts";
 import {
@@ -33,11 +40,7 @@ import {
 } from "./scheduler.ts";
 import { discoverTargets, resolveOrderingEdges } from "./build.ts";
 import { type OrderingEdge, planGraph } from "./graph.ts";
-import {
-  discoverParameters,
-  ParameterError,
-  resolveParameters,
-} from "./params.ts";
+import { ParameterError } from "./params.ts";
 import {
   type BuildCache,
   CACHE_FILE,
@@ -46,13 +49,8 @@ import {
   openCache,
 } from "./cache.ts";
 import { findConfigDir, pathExists } from "./config.ts";
-import {
-  type AffectedOptions,
-  affectedTargets,
-  gitChangedFiles,
-} from "./affected.ts";
+import type { AffectedOptions } from "./affected.ts";
 import { type RemoteCacheStore, resolveRemoteStore } from "./remote_cache.ts";
-import { isCI } from "./host.ts";
 import { ServiceRegistry } from "./service.ts";
 import { absolutePath } from "./path.ts";
 import type { TargetBuilder } from "./target.ts";
@@ -222,63 +220,6 @@ function priorWaitsOf(record: RunRecord): ReadonlyMap<string, WaitState> {
   return waits;
 }
 
-/** Prompt for a missing required parameter, only at an interactive (non-CI) TTY. */
-function defaultPrompt(
-  flag: string,
-  description: string | undefined,
-): string | undefined {
-  let interactive = false;
-  try {
-    interactive = Deno.stdin.isTerminal();
-  } catch {
-    interactive = false;
-  }
-  if (!interactive || isCI()) return undefined;
-  const label = description ? `--${flag} (${description})` : `--${flag}`;
-  return prompt(`${label}:`) ?? undefined;
-}
-
-/**
- * Evaluate up-front conditions for `whenSkipped("skip-dependencies")` targets;
- * return the names to skip — those targets plus any dependencies that no other
- * target in the plan needs.
- */
-async function conditionSkips(
-  root: TargetBuilder,
-  order: TargetBuilder[],
-): Promise<Set<string>> {
-  const pruned = new Set<TargetBuilder>();
-  for (const t of order) {
-    if (!t.skipDependencies_ || t.onlyWhen_.length === 0) continue;
-    let run = true;
-    for (const condition of t.onlyWhen_) {
-      if (!(await condition())) {
-        run = false;
-        break;
-      }
-    }
-    if (!run) pruned.add(t);
-  }
-  if (pruned.size === 0) return new Set();
-
-  // Everything still reachable from the root without pulling dependencies in
-  // *through* a pruned target.
-  const kept = new Set<TargetBuilder>();
-  const walk = (node: TargetBuilder) => {
-    if (node === undefined || kept.has(node)) return;
-    kept.add(node);
-    if (pruned.has(node)) return;
-    for (const dep of node.dependsOn_) walk(dep);
-    for (const trigger of node.triggers_) walk(trigger);
-  };
-  walk(root);
-
-  const names = new Set<string>();
-  for (const t of pruned) names.add(t.name_ ?? "");
-  for (const t of order) if (!kept.has(t)) names.add(t.name_ ?? "");
-  return names;
-}
-
 /**
  * Execute the requested target and its transitive dependencies.
  *
@@ -303,26 +244,14 @@ export async function execute(
   } = composeOutput(options);
   const skip = new Set(options.skip ?? []);
 
-  // Resolve declared parameters (CLI value → environment → default) before any
-  // target runs, so a target body can read `this.param.value`. A missing
-  // required parameter or an invalid value fails the build before it starts.
-  const params = discoverParameters(build);
   const readEnv = options.readEnv ?? defaultReadEnv;
-  const paramErrors = await resolveParameters(
-    params,
+  const { params, errors: paramErrors } = await resolveRunParameters(
+    build,
     options.params ?? {},
     readEnv,
     options.prompt ?? defaultPrompt,
     redactor,
   );
-  // Register each secret's final parsed value too — its raw form was already
-  // added during resolution, but a source that trims or a parser that
-  // normalises could yield a slightly different printed string.
-  for (const p of params.values()) {
-    if (!p.secret_) continue;
-    const value = p.stringValue_();
-    if (value !== undefined && value !== "") redactor.add(value);
-  }
   if (style.github && writesToConsole) {
     const secrets: string[] = [];
     for (const p of params.values()) {
@@ -361,35 +290,7 @@ export async function execute(
     return { ok: false, executed: [], error };
   }
   const { order, predecessors } = planGraph(root, extraEdges);
-  // Flag ordering edges that can never apply: an endpoint that is neither in this
-  // run's execution set nor a declared build target is dead weight (silently
-  // dropped otherwise). A declared target simply not in this run — a conditional
-  // target — is legitimately ignored, so it is not flagged. This catches feeding
-  // ad-hoc or fan-out per-item names into orderWith/extraEdges: per-item fan-out
-  // ordering is not expressible (order whole fan-out waves with .dependsOn).
-  if (extraEdges.length > 0) {
-    const inRun = new Set(order);
-    const declared = new Set(discovered.values());
-    const dangling = new Set<string>();
-    for (const edge of extraEdges) {
-      for (const endpoint of edge) {
-        // `endpoint &&` tolerates a malformed edge (a consumer bypassing the
-        // OrderingEdge type to pass a nullish endpoint) instead of crashing on
-        // `.name_`, matching planEdges' silent tolerance of such edges.
-        if (endpoint && !inRun.has(endpoint) && !declared.has(endpoint)) {
-          dangling.add(endpoint.name_ ?? "<unnamed>");
-        }
-      }
-    }
-    for (const name of dangling) {
-      reporter.info(
-        `ordering: an edge references "${name}", which is not a target in this ` +
-          `build — the edge is ignored. orderWith/extraEdges see only ` +
-          `class-field targets, not fan-out sub-targets (order whole fan-out ` +
-          `waves with .dependsOn instead).`,
-      );
-    }
-  }
+  reportDanglingEdges(extraEdges, order, discovered.values(), reporter);
   // Evaluate up-front conditions for `whenSkipped("skip-dependencies")` targets
   // and skip them plus any dependencies that nothing else needs.
   for (const name of await conditionSkips(root, order)) skip.add(name);
@@ -398,15 +299,7 @@ export async function execute(
   // targets still unblock their dependents (their prior outputs are assumed
   // current), so an affected target downstream of an unaffected one still runs.
   if (options.affected !== undefined) {
-    const base = options.affected.base ?? "HEAD";
-    const changedFiles = options.affected.changedFiles ?? gitChangedFiles;
-    const affected = affectedTargets(order, await changedFiles(base));
-    for (const t of order) {
-      if (!affected.has(t)) skip.add(t.name_ ?? "<unnamed>");
-    }
-    if (affected.size === 0) {
-      reporter.info(`No targets affected by changes since ${base}.`);
-    }
+    await applyAffectedSkips(options.affected, order, skip, reporter);
   }
 
   const limit = resolveConcurrency(options.parallel);
